moving things to src, zane recommended
