+++ src/pilot2_twosignals.cpp
